+++ e_ood/__init__.py
@@ -1,1 +1,6 @@
 __version__ = '0.1.36a1'
+
+from .analyze import Analyzer  # noqa
+from .db import ReportedUpdateTypes, VersionDB  # noqa
+from .pypi import PackageVersionInfo  # noqa
+from .virtualenv import EnvPackages  # noqa