+++ CHANGES.md
@@ -27,7 +27,7 @@
 * Updates to the default package db
 * Fix a bug that could result in the last package listed in the virtualenv
   being reported twice.
-* six is now a required dependency.
+* six is now a required development dependency.
 
 ## Versions 0.1.16 through 0.1.29